Separated EcalSRCondTools into a namespace and an EDAnalyzer
The static member functions were moved to a separate namespace.
This avoids having other code include/link the module.
