+++ SimCalorimetry/EcalSelectiveReadoutProducers/BuildFile.xml
@@ -1,11 +1,6 @@
-<use name="SimCalorimetry/EcalSelectiveReadoutAlgos"/>
-<use name="FWCore/Framework"/>
 <use name="FWCore/ParameterSet"/>
-<use name="FWCore/MessageLogger"/>
-<use name="DataFormats/EcalDigi"/>
-<use name="CondFormats/DataRecord"/>
 <use name="CondFormats/EcalObjects"/>
-<use name="CondCore/DBOutputService"/>
+<use name="FWCore/Utilities"/>
 <export>
   <lib name="1"/>
 </export>
